backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.0
cerebras-cloud-sdk==1.15.0
gspread==6.1.0
google-auth==2.35.0
gradio==4.44.1
huggingface-hub<0.25.0
